1-1/dl/dl2021-ex11-transformers-dl2021-pes/requirements.txt
# Install PyTorch manually as described in the Assignment Sheet.
matplotlib>=3.3
numpy>=1.19
pytest>=6.1
pycodestyle>=2.6
scipy>=1.5
scikit-learn>=0.23
seaborn>=0.11.2
torchvision==0.8.2
regex==2019.3.12
absl-py==0.7.1
